@@ -30,7 +30,7 @@
     "build:core": "ng-packagr -p packages/core/package.json",
     "build:js-marker-clusterer": "ng-packagr -p packages/js-marker-clusterer/package.json",
     "build:snazzy-info-window": "ng-packagr -p packages/snazzy-info-window/package.json",
-    "lint": "tslint -c tslint.json 'packages/**/*.ts'",
+    "lint": "tslint --project tsconfig.json -c tslint.json 'packages/**/*.ts'",
     "packagejson": "node ./scripts/update-package-json.js",
     "clang:format": "clang-format --glob=packages/**/*.ts -i",
     "test": "jest",
@@ -55,17 +55,16 @@
     "@angular/core": "^6.0.0 || ^7.0.0 || ^8.0.0"
   },
   "devDependencies": {
-    "@angular/animations": "^7.0.0",
-    "@angular/common": "^7.0.0",
-    "@angular/compiler": "^7.0.0",
-    "@angular/compiler-cli": "^7.0.0",
-    "@angular/core": "^7.0.0",
-    "@angular/http": "^7.0.0",
-    "@angular/platform-browser": "^7.0.0",
-    "@angular/platform-browser-dynamic": "^7.0.0",
-    "@angular/platform-server": "^7.0.0",
+    "@angular/animations": "^8.0.0",
+    "@angular/common": "^8.0.0",
+    "@angular/compiler": "^8.0.0",
+    "@angular/compiler-cli": "^8.0.0",
+    "@angular/core": "^8.0.0",
+    "@angular/platform-browser": "^8.0.0",
+    "@angular/platform-browser-dynamic": "^8.0.0",
+    "@angular/platform-server": "^8.0.0",
     "@compodoc/compodoc": "^1.0.0-beta.9",
-    "@ngtools/webpack": "^7.2.1",
+    "@types/googlemaps": "^3.36.2",
     "@types/jest": "^23.0.0",
     "angular2-template-loader": "^0.6.2",
     "awesome-typescript-loader": "^5.2.1",
@@ -94,11 +93,12 @@
     "ts-helpers": "^1.1.2",
     "ts-jest": "^23.10.5",
     "ts-loader": "^4.3.0",
+    "tslib": "^1.9.3",
     "tslint": "^5.8.0",
     "tslint-loader": "^3.4.3",
-    "typescript": "~3.2.2",
+    "typescript": "3.4.5",
     "webpack": "4.8.0",
-    "zone.js": "^0.8.26"
+    "zone.js": "~0.9.1"
   },
   "jspm": {
     "jspmNodeConversion": false,
