@@ -1,7 +1,7 @@
+/// <reference types="@types/googlemaps" />
 import { Directive, EventEmitter, OnDestroy, OnInit, OnChanges, SimpleChanges, Input, Output } from '@angular/core';
 import { Subscription } from 'rxjs';
 
-import { DataMouseEvent, DataOptions } from './../services/google-maps-types';
 import { DataLayerManager } from './../services/managers/data-layer-manager';
 
 let layerId = 0;
@@ -200,7 +200,7 @@ let layerId = 0;
   selector: 'agm-data-layer'
 })
 export class AgmDataLayer implements OnInit, OnDestroy, OnChanges {
-  private static _dataOptionsAttributes: Array<string> = ['style'];
+  private static _dataOptionsAttributes: string[] = ['style'];
 
   private _addedToManager: boolean = false;
   private _id: string = (layerId++).toString();
@@ -209,7 +209,7 @@ export class AgmDataLayer implements OnInit, OnDestroy, OnChanges {
   /**
    * This event is fired when a feature in the layer is clicked.
    */
-  @Output() layerClick: EventEmitter<DataMouseEvent> = new EventEmitter<DataMouseEvent>();
+  @Output() layerClick: EventEmitter<google.maps.Data.MouseEvent> = new EventEmitter<google.maps.Data.MouseEvent>();
 
   /**
    * The geoJson to be displayed
@@ -234,7 +234,7 @@ export class AgmDataLayer implements OnInit, OnDestroy, OnChanges {
 
   private _addEventListeners() {
     const listeners = [
-      { name: 'click', handler: (ev: DataMouseEvent) => this.layerClick.emit(ev) },
+      { name: 'click', handler: (ev: google.maps.Data.MouseEvent) => this.layerClick.emit(ev) },
     ];
     listeners.forEach((obj) => {
       const os = this._manager.createEventObservable(obj.name, this).subscribe(obj.handler);
@@ -266,7 +266,7 @@ export class AgmDataLayer implements OnInit, OnDestroy, OnChanges {
       this._manager.updateGeoJson(this, geoJsonChange.currentValue);
     }
 
-    let dataOptions: DataOptions = {};
+    let dataOptions: google.maps.Data.DataOptions = {};
 
     AgmDataLayer._dataOptionsAttributes.forEach(k => (<any>dataOptions)[k] = changes.hasOwnProperty(k) ? changes[k].currentValue : (<any>this)[k]);
 
