@@ -1,13 +1,9 @@
+/// <reference types="@types/googlemaps" />
 import { Component, ElementRef, EventEmitter, OnChanges, OnDestroy, OnInit, SimpleChanges, Input, Output, NgZone } from '@angular/core';
 import {Subscription} from 'rxjs';
 
-import {MouseEvent} from '../map-types';
+import {MapRestriction, MouseEvent} from '../map-types';
 import {GoogleMapsAPIWrapper} from '../services/google-maps-api-wrapper';
-import {
-  FullscreenControlOptions, LatLng, LatLngLiteral, MapTypeControlOptions, MapTypeId, PanControlOptions, MapRestriction,
-  RotateControlOptions, ScaleControlOptions, StreetViewControlOptions, ZoomControlOptions
-} from '../services/google-maps-types';
-import {LatLngBounds, LatLngBoundsLiteral, MapTypeStyle} from '../services/google-maps-types';
 import {CircleManager} from '../services/managers/circle-manager';
 import {RectangleManager} from '../services/managers/rectangle-manager';
 import {InfoWindowManager} from '../services/managers/info-window-manager';
@@ -18,7 +14,7 @@ import {KmlLayerManager} from './../services/managers/kml-layer-manager';
 import {DataLayerManager} from './../services/managers/data-layer-manager';
 import {FitBoundsService} from '../services/fit-bounds';
 
-declare var google: any;
+type MapTypeId = 'roadmap'|'hybrid'|'satellite'|'terrain';
 
 /**
  * AgmMap renders a Google Map.
@@ -156,13 +152,13 @@ export class AgmMap implements OnChanges, OnInit, OnDestroy {
   /**
    * Options for the Zoom control.
    */
-  @Input() zoomControlOptions: ZoomControlOptions;
+  @Input() zoomControlOptions: google.maps.ZoomControlOptions;
 
   /**
    * Styles to apply to each of the default map types. Note that for Satellite/Hybrid and Terrain
    * modes, these styles will only apply to labels and geometry.
    */
-  @Input() styles: MapTypeStyle[] = [];
+  @Input() styles: google.maps.MapTypeStyle[] = [];
 
   /**
    * When true and the latitude and/or longitude values changes, the Google Maps panTo method is
@@ -181,13 +177,13 @@ export class AgmMap implements OnChanges, OnInit, OnDestroy {
   /**
    * Options for the Street View control.
    */
-  @Input() streetViewControlOptions: StreetViewControlOptions;
+  @Input() streetViewControlOptions: google.maps.StreetViewControlOptions;
 
   /**
    * Sets the viewport to contain the given bounds.
    * If this option to `true`, the bounds get automatically computed from all elements that use the {@link AgmFitBounds} directive.
    */
-  @Input() fitBounds: LatLngBoundsLiteral|LatLngBounds|boolean = false;
+  @Input() fitBounds: google.maps.LatLngBoundsLiteral|google.maps.LatLngBounds|boolean = false;
 
   /**
    * The initial enabled/disabled state of the Scale control. This is disabled by default.
@@ -197,7 +193,7 @@ export class AgmMap implements OnChanges, OnInit, OnDestroy {
   /**
    * Options for the scale control.
    */
-  @Input() scaleControlOptions: ScaleControlOptions;
+  @Input() scaleControlOptions: google.maps.ScaleControlOptions;
 
   /**
    * The initial enabled/disabled state of the Map type control.
@@ -207,7 +203,7 @@ export class AgmMap implements OnChanges, OnInit, OnDestroy {
   /**
    * Options for the Map type control.
    */
-  @Input() mapTypeControlOptions: MapTypeControlOptions;
+  @Input() mapTypeControlOptions: google.maps.MapTypeControlOptions;
 
   /**
    * The initial enabled/disabled state of the Pan control.
@@ -217,7 +213,7 @@ export class AgmMap implements OnChanges, OnInit, OnDestroy {
   /**
    * Options for the Pan control.
    */
-  @Input() panControlOptions: PanControlOptions;
+  @Input() panControlOptions: google.maps.PanControlOptions;
 
   /**
    * The initial enabled/disabled state of the Rotate control.
@@ -227,7 +223,7 @@ export class AgmMap implements OnChanges, OnInit, OnDestroy {
   /**
    * Options for the Rotate control.
    */
-  @Input() rotateControlOptions: RotateControlOptions;
+  @Input() rotateControlOptions: google.maps.RotateControlOptions;
 
   /**
    * The initial enabled/disabled state of the Fullscreen control.
@@ -237,12 +233,12 @@ export class AgmMap implements OnChanges, OnInit, OnDestroy {
   /**
    * Options for the Fullscreen control.
    */
-  @Input() fullscreenControlOptions: FullscreenControlOptions;
+  @Input() fullscreenControlOptions: google.maps.FullscreenControlOptions;
 
   /**
    * The map mapTypeId. Defaults to 'roadmap'.
    */
-  @Input() mapTypeId: 'roadmap'|'hybrid'|'satellite'|'terrain'|string = 'roadmap';
+  @Input() mapTypeId: MapTypeId = 'roadmap';
 
   /**
    * When false, map icons are not clickable. A map icon represents a point of interest,
@@ -317,12 +313,12 @@ export class AgmMap implements OnChanges, OnInit, OnDestroy {
   /**
    * This event emitter is fired when the map center changes.
    */
-  @Output() centerChange: EventEmitter<LatLngLiteral> = new EventEmitter<LatLngLiteral>();
+  @Output() centerChange: EventEmitter<google.maps.LatLngLiteral> = new EventEmitter<google.maps.LatLngLiteral>();
 
   /**
    * This event is fired when the viewport bounds have changed.
    */
-  @Output() boundsChange: EventEmitter<LatLngBounds> = new EventEmitter<LatLngBounds>();
+  @Output() boundsChange: EventEmitter<google.maps.LatLngBounds> = new EventEmitter<google.maps.LatLngBounds>();
 
   /**
    * This event is fired when the mapTypeId property changes.
@@ -383,12 +379,12 @@ export class AgmMap implements OnChanges, OnInit, OnDestroy {
       rotateControlOptions: this.rotateControlOptions,
       fullscreenControl: this.fullscreenControl,
       fullscreenControlOptions: this.fullscreenControlOptions,
-      mapTypeId: this.mapTypeId,
+      mapTypeId: (this.mapTypeId as unknown) as google.maps.MapTypeId,
       clickableIcons: this.clickableIcons,
       gestureHandling: this.gestureHandling,
       tilt: this.tilt,
       restriction: this.restriction,
-    })
+    } as google.maps.MapOptions /* TODO: @types/googlemaps MapOptions should include restrictions */)
       .then(() => this._mapsWrapper.getNativeMap())
       .then(map => this.mapReady.emit(map));
 
@@ -502,7 +498,7 @@ export class AgmMap implements OnChanges, OnInit, OnDestroy {
     });
   }
 
-  protected _updateBounds(bounds: LatLngBounds|LatLngBoundsLiteral) {
+  protected _updateBounds(bounds: google.maps.LatLngBounds|google.maps.LatLngBoundsLiteral) {
     if (this._isLatLngBoundsLiteral(bounds) && google && google.maps) {
       const newBounds = new google.maps.LatLngBounds();
       newBounds.union(bounds);
@@ -515,16 +511,16 @@ export class AgmMap implements OnChanges, OnInit, OnDestroy {
     this._mapsWrapper.fitBounds(bounds);
   }
 
-  private _isLatLngBoundsLiteral(bounds: LatLngBounds|LatLngBoundsLiteral): bounds is LatLngBoundsLiteral {
+  private _isLatLngBoundsLiteral(bounds: google.maps.LatLngBounds|google.maps.LatLngBoundsLiteral): bounds is google.maps.LatLngBoundsLiteral {
     return bounds != null && (<any>bounds).extend === undefined;
   }
 
   private _handleMapCenterChange() {
     const s = this._mapsWrapper.subscribeToMapEvent<void>('center_changed').subscribe(() => {
-      this._mapsWrapper.getCenter().then((center: LatLng) => {
+      this._mapsWrapper.getCenter().then((center: google.maps.LatLng) => {
         this.latitude = center.lat();
         this.longitude = center.lng();
-        this.centerChange.emit(<LatLngLiteral>{lat: this.latitude, lng: this.longitude});
+        this.centerChange.emit(<google.maps.LatLngLiteral>{lat: this.latitude, lng: this.longitude});
       });
     });
     this._observableSubscriptions.push(s);
@@ -533,7 +529,7 @@ export class AgmMap implements OnChanges, OnInit, OnDestroy {
   private _handleBoundsChange() {
     const s = this._mapsWrapper.subscribeToMapEvent<void>('bounds_changed').subscribe(() => {
       this._mapsWrapper.getBounds().then(
-          (bounds: LatLngBounds) => { this.boundsChange.emit(bounds); });
+          (bounds: google.maps.LatLngBounds) => { this.boundsChange.emit(bounds); });
     });
     this._observableSubscriptions.push(s);
   }
@@ -541,7 +537,7 @@ export class AgmMap implements OnChanges, OnInit, OnDestroy {
   private _handleMapTypeIdChange() {
     const s = this._mapsWrapper.subscribeToMapEvent<void>('maptypeid_changed').subscribe(() => {
       this._mapsWrapper.getMapTypeId().then(
-          (mapTypeId: MapTypeId) => { this.mapTypeIdChange.emit(mapTypeId); });
+          (mapTypeId: string) => { this.mapTypeIdChange.emit(mapTypeId as MapTypeId); });
     });
     this._observableSubscriptions.push(s);
   }
@@ -575,8 +571,8 @@ export class AgmMap implements OnChanges, OnInit, OnDestroy {
     ];
 
     events.forEach((e: Event) => {
-      const s = this._mapsWrapper.subscribeToMapEvent<{latLng: LatLng}>(e.name).subscribe(
-          (event: {latLng: LatLng}) => {
+      const s = this._mapsWrapper.subscribeToMapEvent<{latLng: google.maps.LatLng}>(e.name).subscribe(
+          (event: {latLng: google.maps.LatLng}) => {
             const value = <MouseEvent>{coords: {lat: event.latLng.lat(), lng: event.latLng.lng()}};
             e.emitter.emit(value);
           });
