@@ -1,8 +1,9 @@
+/// <reference types="@types/googlemaps" />
 import { AfterContentInit, ContentChildren, Directive, EventEmitter, OnChanges, OnDestroy, QueryList, SimpleChanges, Input, Output } from '@angular/core';
 import { Subscription } from 'rxjs';
 
-import { PolyMouseEvent, LatLng } from '../services/google-maps-types';
 import { PolylineManager } from '../services/managers/polyline-manager';
+import { MVCEvent } from '../utils/mvcarray-utils';
 import { AgmPolylinePoint } from './polyline-point';
 
 let polylineId = 0;
@@ -91,12 +92,12 @@ export class AgmPolyline implements OnDestroy, OnChanges, AfterContentInit {
   /**
    * This event is fired when the DOM click event is fired on the Polyline.
    */
-  @Output() lineClick: EventEmitter<PolyMouseEvent> = new EventEmitter<PolyMouseEvent>();
+  @Output() lineClick: EventEmitter<google.maps.PolyMouseEvent> = new EventEmitter<google.maps.PolyMouseEvent>();
 
   /**
    * This event is fired when the DOM dblclick event is fired on the Polyline.
    */
-  @Output() lineDblClick: EventEmitter<PolyMouseEvent> = new EventEmitter<PolyMouseEvent>();
+  @Output() lineDblClick: EventEmitter<google.maps.PolyMouseEvent> = new EventEmitter<google.maps.PolyMouseEvent>();
 
   /**
    * This event is repeatedly fired while the user drags the polyline.
@@ -116,32 +117,32 @@ export class AgmPolyline implements OnDestroy, OnChanges, AfterContentInit {
   /**
    * This event is fired when the DOM mousedown event is fired on the Polyline.
    */
-  @Output() lineMouseDown: EventEmitter<PolyMouseEvent> = new EventEmitter<PolyMouseEvent>();
+  @Output() lineMouseDown: EventEmitter<google.maps.PolyMouseEvent> = new EventEmitter<google.maps.PolyMouseEvent>();
 
   /**
    * This event is fired when the DOM mousemove event is fired on the Polyline.
    */
-  @Output() lineMouseMove: EventEmitter<PolyMouseEvent> = new EventEmitter<PolyMouseEvent>();
+  @Output() lineMouseMove: EventEmitter<google.maps.PolyMouseEvent> = new EventEmitter<google.maps.PolyMouseEvent>();
 
   /**
    * This event is fired on Polyline mouseout.
    */
-  @Output() lineMouseOut: EventEmitter<PolyMouseEvent> = new EventEmitter<PolyMouseEvent>();
+  @Output() lineMouseOut: EventEmitter<google.maps.PolyMouseEvent> = new EventEmitter<google.maps.PolyMouseEvent>();
 
   /**
    * This event is fired on Polyline mouseover.
    */
-  @Output() lineMouseOver: EventEmitter<PolyMouseEvent> = new EventEmitter<PolyMouseEvent>();
+  @Output() lineMouseOver: EventEmitter<google.maps.PolyMouseEvent> = new EventEmitter<google.maps.PolyMouseEvent>();
 
   /**
    * This event is fired whe the DOM mouseup event is fired on the Polyline
    */
-  @Output() lineMouseUp: EventEmitter<PolyMouseEvent> = new EventEmitter<PolyMouseEvent>();
+  @Output() lineMouseUp: EventEmitter<google.maps.PolyMouseEvent> = new EventEmitter<google.maps.PolyMouseEvent>();
 
   /**
    * This event is fired when the Polyline is right-clicked on.
    */
-  @Output() lineRightClick: EventEmitter<PolyMouseEvent> = new EventEmitter<PolyMouseEvent>();
+  @Output() lineRightClick: EventEmitter<google.maps.PolyMouseEvent> = new EventEmitter<google.maps.PolyMouseEvent>();
 
   /**
    * This event is fired after Polyline's path changes.
@@ -153,7 +154,7 @@ export class AgmPolyline implements OnDestroy, OnChanges, AfterContentInit {
    */
   @ContentChildren(AgmPolylinePoint) points: QueryList<AgmPolylinePoint>;
 
-  private static _polylineOptionsAttributes: Array<string> = [
+  private static _polylineOptionsAttributes: string[] = [
     'draggable', 'editable', 'visible', 'geodesic', 'strokeColor', 'strokeOpacity', 'strokeWeight',
     'zIndex'
   ];
@@ -194,7 +195,7 @@ export class AgmPolyline implements OnDestroy, OnChanges, AfterContentInit {
     this._polylineManager.setPolylineOptions(this, options);
   }
 
-  getPath(): Promise<Array<LatLng>> {
+  getPath(): Promise<google.maps.LatLng[]> {
     return this._polylineManager.getPath(this);
   }
 
@@ -206,17 +207,17 @@ export class AgmPolyline implements OnDestroy, OnChanges, AfterContentInit {
 
   private _addEventListeners() {
     const handlers = [
-      {name: 'click', handler: (ev: PolyMouseEvent) => this.lineClick.emit(ev)},
-      {name: 'dblclick', handler: (ev: PolyMouseEvent) => this.lineDblClick.emit(ev)},
+      {name: 'click', handler: (ev: google.maps.PolyMouseEvent) => this.lineClick.emit(ev)},
+      {name: 'dblclick', handler: (ev: google.maps.PolyMouseEvent) => this.lineDblClick.emit(ev)},
       {name: 'drag', handler: (ev: MouseEvent) => this.lineDrag.emit(ev)},
       {name: 'dragend', handler: (ev: MouseEvent) => this.lineDragEnd.emit(ev)},
       {name: 'dragstart', handler: (ev: MouseEvent) => this.lineDragStart.emit(ev)},
-      {name: 'mousedown', handler: (ev: PolyMouseEvent) => this.lineMouseDown.emit(ev)},
-      {name: 'mousemove', handler: (ev: PolyMouseEvent) => this.lineMouseMove.emit(ev)},
-      {name: 'mouseout', handler: (ev: PolyMouseEvent) => this.lineMouseOut.emit(ev)},
-      {name: 'mouseover', handler: (ev: PolyMouseEvent) => this.lineMouseOver.emit(ev)},
-      {name: 'mouseup', handler: (ev: PolyMouseEvent) => this.lineMouseUp.emit(ev)},
-      {name: 'rightclick', handler: (ev: PolyMouseEvent) => this.lineRightClick.emit(ev)},
+      {name: 'mousedown', handler: (ev: google.maps.PolyMouseEvent) => this.lineMouseDown.emit(ev)},
+      {name: 'mousemove', handler: (ev: google.maps.PolyMouseEvent) => this.lineMouseMove.emit(ev)},
+      {name: 'mouseout', handler: (ev: google.maps.PolyMouseEvent) => this.lineMouseOut.emit(ev)},
+      {name: 'mouseover', handler: (ev: google.maps.PolyMouseEvent) => this.lineMouseOver.emit(ev)},
+      {name: 'mouseup', handler: (ev: google.maps.PolyMouseEvent) => this.lineMouseUp.emit(ev)},
+      {name: 'rightclick', handler: (ev: google.maps.PolyMouseEvent) => this.lineRightClick.emit(ev)},
     ];
     handlers.forEach((obj) => {
       const os = this._polylineManager.createEventObservable(obj.name, this).subscribe(obj.handler);
@@ -230,7 +231,7 @@ export class AgmPolyline implements OnDestroy, OnChanges, AfterContentInit {
   }
 
   /** @internal */
-  _getPoints(): Array<AgmPolylinePoint> {
+  _getPoints(): AgmPolylinePoint[] {
     if (this.points) {
       return this.points.toArray();
     }
@@ -248,9 +249,5 @@ export class AgmPolyline implements OnDestroy, OnChanges, AfterContentInit {
   }
 }
 
-export interface PathEvent {
-  newArr: LatLng[];
-  evName: 'insert_at' | 'remove_at' | 'set_at';
-  index: number;
-  previous?: LatLng;
+export interface PathEvent extends MVCEvent<google.maps.LatLng> {
 }
