@@ -1,10 +1,8 @@
+/// <reference types="@types/googlemaps" />
 // main modules
 export * from './directives';
 export * from './services';
 export * from './map-types';
-
-// Google Maps types
-export {LatLngBounds, LatLng, LatLngLiteral, MapTypeStyle, PolyMouseEvent} from './services/google-maps-types';
 
 // core module
 // we explicitly export the module here to prevent this Ionic 2 bug:
