@@ -1,17 +1,15 @@
+/// <reference types="@types/googlemaps" />
 import {Injectable, NgZone} from '@angular/core';
 import {Observable, Observer} from 'rxjs';
 
 import {AgmMarker} from './../../directives/marker';
 
 import {GoogleMapsAPIWrapper} from './../google-maps-api-wrapper';
-import {Marker} from './../google-maps-types';
-
-declare var google: any;
 
 @Injectable()
 export class MarkerManager {
-  protected _markers: Map<AgmMarker, Promise<Marker>> =
-      new Map<AgmMarker, Promise<Marker>>();
+  protected _markers: Map<AgmMarker, Promise<google.maps.Marker>> =
+      new Map<AgmMarker, Promise<google.maps.Marker>>();
 
   constructor(protected _mapsWrapper: GoogleMapsAPIWrapper, protected _zone: NgZone) {}
 
@@ -21,7 +19,7 @@ export class MarkerManager {
       // marker already deleted
       return Promise.resolve();
     }
-    return m.then((m: Marker) => {
+    return m.then((m: google.maps.Marker) => {
       return this._zone.run(() => {
         m.setMap(null);
         this._markers.delete(marker);
@@ -31,43 +29,43 @@ export class MarkerManager {
 
   updateMarkerPosition(marker: AgmMarker): Promise<void> {
     return this._markers.get(marker).then(
-        (m: Marker) => m.setPosition({lat: marker.latitude, lng: marker.longitude}));
+        (m: google.maps.Marker) => m.setPosition({lat: marker.latitude, lng: marker.longitude}));
   }
 
   updateTitle(marker: AgmMarker): Promise<void> {
-    return this._markers.get(marker).then((m: Marker) => m.setTitle(marker.title));
+    return this._markers.get(marker).then((m: google.maps.Marker) => m.setTitle(marker.title));
   }
 
   updateLabel(marker: AgmMarker): Promise<void> {
-    return this._markers.get(marker).then((m: Marker) => { m.setLabel(marker.label); });
+    return this._markers.get(marker).then((m: google.maps.Marker) => { m.setLabel(marker.label); });
   }
 
   updateDraggable(marker: AgmMarker): Promise<void> {
-    return this._markers.get(marker).then((m: Marker) => m.setDraggable(marker.draggable));
+    return this._markers.get(marker).then((m: google.maps.Marker) => m.setDraggable(marker.draggable));
   }
 
   updateIcon(marker: AgmMarker): Promise<void> {
-    return this._markers.get(marker).then((m: Marker) => m.setIcon(marker.iconUrl));
+    return this._markers.get(marker).then((m: google.maps.Marker) => m.setIcon(marker.iconUrl));
   }
 
   updateOpacity(marker: AgmMarker): Promise<void> {
-    return this._markers.get(marker).then((m: Marker) => m.setOpacity(marker.opacity));
+    return this._markers.get(marker).then((m: google.maps.Marker) => m.setOpacity(marker.opacity));
   }
 
   updateVisible(marker: AgmMarker): Promise<void> {
-    return this._markers.get(marker).then((m: Marker) => m.setVisible(marker.visible));
+    return this._markers.get(marker).then((m: google.maps.Marker) => m.setVisible(marker.visible));
   }
 
   updateZIndex(marker: AgmMarker): Promise<void> {
-    return this._markers.get(marker).then((m: Marker) => m.setZIndex(marker.zIndex));
+    return this._markers.get(marker).then((m: google.maps.Marker) => m.setZIndex(marker.zIndex));
   }
 
   updateClickable(marker: AgmMarker): Promise<void> {
-    return this._markers.get(marker).then((m: Marker) => m.setClickable(marker.clickable));
+    return this._markers.get(marker).then((m: google.maps.Marker) => m.setClickable(marker.clickable));
   }
 
   updateAnimation(marker: AgmMarker): Promise<void> {
-    return this._markers.get(marker).then((m: Marker) => {
+    return this._markers.get(marker).then((m: google.maps.Marker) => {
       if (typeof marker.animation === 'string') {
         m.setAnimation(google.maps.Animation[marker.animation]);
       } else {
@@ -93,13 +91,13 @@ export class MarkerManager {
     this._markers.set(marker, markerPromise);
   }
 
-  getNativeMarker(marker: AgmMarker): Promise<Marker> {
+  getNativeMarker(marker: AgmMarker): Promise<google.maps.Marker> {
     return this._markers.get(marker);
   }
 
   createEventObservable<T>(eventName: string, marker: AgmMarker): Observable<T> {
     return new Observable((observer: Observer<T>) => {
-      this._markers.get(marker).then((m: Marker) => {
+      this._markers.get(marker).then((m: google.maps.Marker) => {
         m.addListener(eventName, (e: T) => this._zone.run(() => observer.next(e)));
       });
     });
