@@ -1,16 +1,16 @@
+/// <reference types="@types/googlemaps" />
 import { Injectable, NgZone } from '@angular/core';
 import { Observable, Observer, merge } from 'rxjs';
 import { startWith, map, switchMap, skip } from 'rxjs/operators';
 
 import {AgmPolygon, PolygonPathEvent, PathCollectionChangePolygonPathEvent, PathChangePolygonPathEvent} from '../../directives/polygon';
 import {GoogleMapsAPIWrapper} from '../google-maps-api-wrapper';
-import {Polygon, LatLng, MVCArray} from '../google-maps-types';
 import { createMVCEventObservable, MVCEvent } from '../../utils/mvcarray-utils';
 
 @Injectable()
 export class PolygonManager {
-  private _polygons: Map<AgmPolygon, Promise<Polygon>> =
-    new Map<AgmPolygon, Promise<Polygon>>();
+  private _polygons: Map<AgmPolygon, Promise<google.maps.Polygon>> =
+    new Map<AgmPolygon, Promise<google.maps.Polygon>>();
 
   constructor(private _mapsWrapper: GoogleMapsAPIWrapper, private _zone: NgZone) { }
 
@@ -37,11 +37,11 @@ export class PolygonManager {
     if (m == null) {
       return Promise.resolve();
     }
-    return m.then((l: Polygon) => this._zone.run(() => { l.setPaths(polygon.paths); }));
+    return m.then((l: google.maps.Polygon) => this._zone.run(() => { l.setPaths(polygon.paths); }));
   }
 
   setPolygonOptions(path: AgmPolygon, options: { [propName: string]: any }): Promise<void> {
-    return this._polygons.get(path).then((l: Polygon) => { l.setOptions(options); });
+    return this._polygons.get(path).then((l: google.maps.Polygon) => { l.setOptions(options); });
   }
 
   deletePolygon(paths: AgmPolygon): Promise<void> {
@@ -49,7 +49,7 @@ export class PolygonManager {
     if (m == null) {
       return Promise.resolve();
     }
-    return m.then((l: Polygon) => {
+    return m.then((l: google.maps.Polygon) => {
       return this._zone.run(() => {
         l.setMap(null);
         this._polygons.delete(paths);
@@ -57,29 +57,29 @@ export class PolygonManager {
     });
   }
 
-  getPath(polygon: AgmPolygon): Promise<Array<LatLng>> {
+  getPath(polygon: AgmPolygon): Promise<google.maps.LatLng[]> {
     return this._polygons.get(polygon)
       .then((polygon) => polygon.getPath().getArray());
   }
 
-  getPaths(polygon: AgmPolygon): Promise<Array<Array<LatLng>>> {
+  getPaths(polygon: AgmPolygon): Promise<google.maps.LatLng[][]> {
     return this._polygons.get(polygon)
       .then((polygon) => polygon.getPaths().getArray().map((p) => p.getArray()));
   }
 
   createEventObservable<T>(eventName: string, path: AgmPolygon): Observable<T> {
     return new Observable((observer: Observer<T>) => {
-      this._polygons.get(path).then((l: Polygon) => {
+      this._polygons.get(path).then((l: google.maps.Polygon) => {
         l.addListener(eventName, (e: T) => this._zone.run(() => observer.next(e)));
       });
     });
   }
 
-  async createPathEventObservable(agmPolygon: AgmPolygon): Promise<Observable<PolygonPathEvent<any>>> {
+  async createPathEventObservable(agmPolygon: AgmPolygon): Promise<Observable<PolygonPathEvent>> {
     const polygon = await this._polygons.get(agmPolygon);
     const paths = polygon.getPaths();
     const pathsChanges$ = createMVCEventObservable(paths);
-    return pathsChanges$.pipe(startWith(({ newArr: paths.getArray() } as MVCEvent<MVCArray<LatLng>>)), // in order to subscribe to them all
+    return pathsChanges$.pipe(startWith(({ newArr: paths.getArray() } as MVCEvent<google.maps.MVCArray<google.maps.LatLng>>)), // in order to subscribe to them all
       switchMap(parentMVEvent => merge(... // rest parameter
         parentMVEvent.newArr.map((chMVC, index) =>
           createMVCEventObservable(chMVC)
@@ -91,7 +91,7 @@ export class PolygonManager {
         if (!chMVCEvent) {
           retVal = {
             newArr: parentMVEvent.newArr.map(subArr => subArr.getArray().map(latLng => latLng.toJSON())),
-            eventName: parentMVEvent.evName,
+            eventName: parentMVEvent.eventName,
             index: parentMVEvent.index,
           } as PathCollectionChangePolygonPathEvent;
           if (parentMVEvent.previous) {
@@ -101,7 +101,7 @@ export class PolygonManager {
           retVal = {
             newArr: parentMVEvent.newArr.map(subArr => subArr.getArray().map(latLng => latLng.toJSON())),
             pathIndex,
-            eventName: chMVCEvent.evName,
+            eventName: chMVCEvent.eventName,
             index: chMVCEvent.index
           } as PathChangePolygonPathEvent;
           if (chMVCEvent.previous) {
