@@ -1,9 +1,9 @@
+/// <reference types="@types/googlemaps" />
 import {NgZone} from '@angular/core';
 import {TestBed, inject} from '@angular/core/testing';
 
 import {AgmPolyline} from '../../directives/polyline';
 import {GoogleMapsAPIWrapper} from '../../services/google-maps-api-wrapper';
-import {Polyline} from '../../services/google-maps-types';
 import {PolylineManager} from '../../services/managers/polyline-manager';
 
 describe('PolylineManager', () => {
@@ -51,7 +51,7 @@ describe('PolylineManager', () => {
            (polylineManager: PolylineManager, apiWrapper: GoogleMapsAPIWrapper) => {
              const newPolyline = new AgmPolyline(polylineManager);
 
-             const polylineInstance: Partial<Polyline> = {
+             const polylineInstance: Partial<google.maps.Polyline> = {
               setMap: jest.fn()
              };
              (<jest.Mock>apiWrapper.createPolyline).mockReturnValue(Promise.resolve(polylineInstance));
