@@ -1,22 +1,22 @@
+/// <reference types="@types/googlemaps" />
 import {Injectable, NgZone} from '@angular/core';
 import {Observable, Observer} from 'rxjs';
 
 import {AgmPolyline, PathEvent} from '../../directives/polyline';
 import {AgmPolylinePoint} from '../../directives/polyline-point';
 import {GoogleMapsAPIWrapper} from '../google-maps-api-wrapper';
-import {LatLng, LatLngLiteral, Polyline, MVCArray } from '../google-maps-types';
-import { MVCEvent, createMVCEventObservable } from '../../utils/mvcarray-utils';
+import { createMVCEventObservable } from '../../utils/mvcarray-utils';
 
 @Injectable()
 export class PolylineManager {
-  private _polylines: Map<AgmPolyline, Promise<Polyline>> =
-      new Map<AgmPolyline, Promise<Polyline>>();
+  private _polylines: Map<AgmPolyline, Promise<google.maps.Polyline>> =
+      new Map<AgmPolyline, Promise<google.maps.Polyline>>();
 
   constructor(private _mapsWrapper: GoogleMapsAPIWrapper, private _zone: NgZone) {}
 
-  private static _convertPoints(line: AgmPolyline): Array<LatLngLiteral> {
+  private static _convertPoints(line: AgmPolyline): google.maps.LatLngLiteral[] {
     const path = line._getPoints().map((point: AgmPolylinePoint) => {
-      return <LatLngLiteral>{lat: point.latitude, lng: point.longitude};
+      return <google.maps.LatLngLiteral>{lat: point.latitude, lng: point.longitude};
     });
     return path;
   }
@@ -44,12 +44,12 @@ export class PolylineManager {
     if (m == null) {
       return Promise.resolve();
     }
-    return m.then((l: Polyline) => { return this._zone.run(() => { l.setPath(path); }); });
+    return m.then((l: google.maps.Polyline) => { return this._zone.run(() => { l.setPath(path); }); });
   }
 
   setPolylineOptions(line: AgmPolyline, options: {[propName: string]: any}):
       Promise<void> {
-    return this._polylines.get(line).then((l: Polyline) => { l.setOptions(options); });
+    return this._polylines.get(line).then((l: google.maps.Polyline) => { l.setOptions(options); });
   }
 
   deletePolyline(line: AgmPolyline): Promise<void> {
@@ -57,7 +57,7 @@ export class PolylineManager {
     if (m == null) {
       return Promise.resolve();
     }
-    return m.then((l: Polyline) => {
+    return m.then((l: google.maps.Polyline) => {
       return this._zone.run(() => {
         l.setMap(null);
         this._polylines.delete(line);
@@ -65,18 +65,18 @@ export class PolylineManager {
     });
   }
 
-  private async getMVCPath(agmPolyline: AgmPolyline): Promise<MVCArray<LatLng>> {
+  private async getMVCPath(agmPolyline: AgmPolyline): Promise<google.maps.MVCArray<google.maps.LatLng>> {
     const polyline = await this._polylines.get(agmPolyline);
     return polyline.getPath();
   }
 
-  async getPath(agmPolyline: AgmPolyline): Promise<Array<LatLng>> {
+  async getPath(agmPolyline: AgmPolyline): Promise<google.maps.LatLng[]> {
     return (await this.getMVCPath(agmPolyline)).getArray();
   }
 
   createEventObservable<T>(eventName: string, line: AgmPolyline): Observable<T> {
     return new Observable((observer: Observer<T>) => {
-      this._polylines.get(line).then((l: Polyline) => {
+      this._polylines.get(line).then((l: google.maps.Polyline) => {
         l.addListener(eventName, (e: T) => this._zone.run(() => observer.next(e)));
       });
     });
