@@ -153,7 +153,7 @@ export class LazyMapsAPILoader extends MapsAPILoader {
     }
 
     const hostAndPath: string = this._config.hostAndPath || 'maps.googleapis.com/maps/api/js';
-    const queryParams: {[key: string]: string | Array<string>} = {
+    const queryParams: {[key: string]: string | string[]} = {
       v: this._config.apiVersion || 'quarterly',
       callback: callbackName,
       key: this._config.apiKey,
