@@ -1,9 +1,9 @@
+/// <reference types="@types/googlemaps" />
 import {Injectable, NgZone} from '@angular/core';
 
 import 'js-marker-clusterer';
 
 import {AgmMarker, GoogleMapsAPIWrapper, MarkerManager} from '@agm/core';
-import {Marker} from '@agm/core/services/google-maps-types';
 import {AgmMarkerCluster} from '../../directives/marker-cluster';
 import {MarkerClustererInstance, ClusterOptions} from '../google-clusterer-types';
 
@@ -64,7 +64,7 @@ export class ClusterManager extends MarkerManager {
       // marker already deleted
       return Promise.resolve();
     }
-    return m.then((m: Marker) => {
+    return m.then((m: google.maps.Marker) => {
       this._zone.run(() => {
         m.setMap(null);
         this.getClustererInstance().then(cluster => {
